requirements.txt
streamlit
numpy
pandas
opencv-python
scikit-learn
matplotlib
Pillow
"opencv-python-headless" 
